Continuing work on world map
Click once to move on to an object like planet, startsystem, galaxy or asteroid field. When ON object, click AGAIN to enter. This is probably not the functionality we are looking for, but for now it works well enough Asteroid scene and planet scene can be exited with M.

Stacking scenes: world map is saved, because it is never unloaded. Scenes are just loaded on top of it. And this works quite fine, except there was a little flash of an atrifact during the load. This seems to be gotten rid of by putting another camera to draw blank during the scene transition. Maybe clearing some render buffer or something could work too, but at least now there's some camera rendering all the time.

Some entering functionality to placeholder planet scene and to asteroid field. Asteroid fields are now clickable like planets which was a little tricky to figure out and implement. Asteroid fields changed to use a plane as a collider. This works better than the original sphere. Planet orbits didn't show correctly (wrong color) in the build but things worked fine in the editor. Fixed this by changing the material before drawing and setting a reference to the material on UniverseController.

Some changes to galaxy and universe initializing.