Save load system saves and loads the universe
Now bug free and data about galaxies, star systems, planets and asteroid field is passed back and forth, here and there!

No buttons yet (should we have a new game and load game scene?), we only check if a save file exists and go on with that info.

Things work, but maybe some of the star system initialization could happen in different place. That is, when star system becomes focused. But things seem to work now, so maybe it's not an issue.

Data about planet and so forth is passed to planet launcher. Preventing entering the sun, lol.

The beginnings of the fuel system.