Worm holes and fuel system in place
Wormholes made part of the save system like the rest of universe object. If not enough fuel to any star after going to wormhole, the ship goes to the closest star (presumably running on fumes), because otherwise the game would get stuck and that wouldn't be fun. Rather: game breaking.