Weight limit added to game
Some creative programming around the multiple GameObject.Find -calls that prevent inventory from functioning if the UI is not active. Activating and disabling inventory every time an object is picked up seemed to be the least problematic way of doing this functionality and actually the only way. Otherwise trying to replace finding and references only resulted in annoying NullReferenceExceptions. UI and inventory should definitely be uncoupled from each other...