Life support system for asteroids added and diamond tipped drill added
Life support system added with UI prompts about having to have oxygen bottles and spacesuit on and HUD for oxygen. This whole system works with the inventory. Hiding the prompt and spaceship HUD if the inventory is visible, so we don't draw them on top of inventory.

Diamond drill is just a diamond drill.