Fuel system combined with the inventory + a bug fix and a quality of life fix
DONE: Fuel system combine with inventory
-make fuel system work based on how much fuels we have in inventory

DONE: bug: two planet scenes are loaded on top of another

DONE: quality of life: On world map it should be possible to instantly click a galaxy or star when we are on top of one. Now it requires two clicks and this is not good.