(Saving and loading should be finished for now and bug-free. See message for details. Added icons for valuable stones and replaced shop placeholder with a new model.
SAving and loading pretty much every value that is changed during the game. Inventory, amount of fuels, carbon and water bottles as well as details about hydroponics bay, is it running or not and how long it has been out of resources.

Fixed a bug where player and vendor items could be added twice, now the lists are cleared properly. Every vendor has some fuel and oxygen for now to allow progress on new game. Changed Enter planet method on GameManager, because it used copy paste. Now the strings are set once as well as the indexes